Rename ControlListPage.vue to ControlPage.vue (single control now)
