+++ static/src/controls/ControlPage.vue
